+++ tests/testthat/test-DfReadDataFile3.R
@@ -13,5 +13,24 @@ test_that(contextStr, {
   }
   
   ds <- readRDS(fn)
-  expect_equal(DfReadDataFile(fileName, newExcelFileFormat = TRUE), ds)
+  temp <- DfReadDataFile(fileName, newExcelFileFormat = TRUE)
+  expect_equal(temp, ds)
+  
+  # calculate pseudovalues for split plot A
+  fileName <- system.file(
+    "extdata", "/toyFiles/FROC/frocSpA.xlsx", package = "RJafroc", mustWork = TRUE)
+  temp <- DfReadDataFile(fileName, newExcelFileFormat = TRUE)
+  
+  fn <- paste0(test_path(), "/goodValues361/DfReadDataFile/1T3Rvs4R", ".rds")
+  if (!file.exists(fn)) {
+    warning(paste0("File not found - generating new ",fn))
+    x1 <- UtilPseudoValues(temp, FOM = "wAFROC")
+    saveRDS(temp, file = fn)
+  }
+  
+  x1 <- readRDS(fn)
+  x2 <- UtilPseudoValues(temp, FOM = "wAFROC")
+  expect_equal(x1, x2)
+  
+  
 })